rdp-pipewire-stream: Make use of new rendering path when possible
The new render handling is there. The debug option for the new path is
there. The last missing bit missing for now is the part, where the new
path is used. This happens in the RDP PipeWire stream class.
If the previously added debug option for the new render handling is
activated, make use of the new rendering path.

Do this by simply submitting PipeWire buffers directly to the surface
renderer, when the debug option is activated.