rdp-render-context: Add APIs to acquire and release image views
When trying to render a new frame, at least one image view is needed. In
theory, these could be created on demand, but this would take time.
It is performance-wise more efficient to allocate a set of image views
initially and then use those on demand.
The image views will be created by the encode session. However, the
encode session won't manage which frame uses which image view. This task
is better be done by the render context, as that will allow
gnome-remote-desktop later to progressively encode frames. The encode
session won't know what is supposed to happen on the client's screen.
Its task is just to encode frames.