rdp-surface-renderer: Add getter to get PW buffer info
The render context is the entity, that will create the encode session
and view creator. To be able to determine, whether VAAPI with Vulkan can
be used, it needs to know some information, for example, which buffer
type is used (dma-buf or mem-fd), or whether an explicit modifier is
used.