vulkan: Add shader for stereo view creation
H.264 by default uses subsampled chroma data for its source surfaces
(AVC420). While for video content AVC420 is just fine, as the human eye
won't notice a difference there, with coloured text content, this is
different. With coloured text, the text looks blurry, as only one
chroma-U and one chroma-V value per 2x2 grid is submitted.

The solution for this problem is AVC444. However, not all encoders and
decoders support that. When H.264 was introduced to RDP, (hardware-)
encoders, that did support H.264, only supported AVC420, but not AVC444
(natively). Today, lots of (hardware-) encoders support AVC444 natively,
but still not all, for example AMD GPUs, which today still only support
subsampled video formats.

To solve the blurry text issue with H.264, Microsoft introduced in RDP a
different way of AVC444, that uses an AVC420 stream, but still
transports high resolution colour data.
This technique uses two subframes for one regular frame. The first frame
is a regular YUV420 frame with filtered chroma data. This is called the
main view. The other chroma data is transported via the second frame,
the auxiliary frame.
When the RDP client decodes both subframes, it will reassemble the
original frame from both subframes. This gets rid of the blurry text
issue.

The way, these two subframes are constructed, is documented in 3.3.8.3.3
YUV420p Stream Combination for YUV444v2 mode ([MS-RDPEGFX]). Add a
shader, that creates those views, in order to be able to add support for
AVC444.