rdp: Replace main view shader usage with stereo view shader
Now, in order to be able create the auxiliary view too, replace the
current main view shader usage with the stereo view shader. This mostly
involves binding the auxiliary view to its descriptor set, like it is
done for the main view, but also exchanging the main view shader with
the stereo view shader itself.