rdp-view-creator-avc: Make damage buffer creation function reusable
Currently, there is one state buffer used in the stereo view shader:
the damage buffer. However, there will be another state buffer in the
next commits, a "chroma check buffer". The data in this buffer
represents, whether a tile needs the auxiliary view or not. This allows
gnome-remote-desktop and connected RDP clients later to save bandwidth
and computational power, that would be needed in order to process the
auxiliary view, when that view is actually not needed.

In order to be able to easily create this second buffer, make the damage
buffer creation function reusable.