rdp: Also check whether auxiliary view is actually needed
Not every frame requires an auxiliary view. This is especially the case,
when watching some usual video. In that case, gnome-remote-desktop and
its connected RDP client can save both bandwidth and computational power
by just not processing the auxiliary view.

In order to check whether the auxiliary view is needed, use the cutoff
threshold check from 3.3.8.3.3 YUV420p Stream Combination for YUV444v2
mode ([MS-RDPEGFX]). While this check is intended for the client side
when applying the reverse filter, it can also serve the server side
here, when checking, whether the actual chroma values are too different
from the filtered chroma value.

For now, only perform the auxiliary view check in the shader, but don't
process the result that. The processing part will happen in later
commits.