Add image view for NV12 images
For AVC420, an NV12 surface is needed. An NV12 surface contains two
layers:
One Y-layer for the luma data, and one UV-layer for the subsampled
chroma data. In RDPGFX, this is called the main view.
For AVC444, two NV12 surfaces will be needed: One for the main view and
one for the auxiliary view.
The reason for this is, that in RDPGFX, one AVC420 stream is used for
AVC444. The auxiliary view here contains the missing chroma data from
the main view. After decoding both views (subframes), the client
reassembles the actual frame. Microsoft added this way to RDP, as at the
time of creation, most hardware encoders did not support native AVC444.
AMD GPUs today still don't support native AVC444, but just AVC420.

The image view here in gnome-remote-desktop will be used for each of
those views.
Theoretically, the NV12 image view could just contain one VkImage
containing both the Y- and UV-layer. However, the Intel driver does not
support multiple image layers in Vulkan, although Vulkan supports that.
So use for each layers an individual VkImage.