rdp-pipewire-stream: Remove unused CUDA handling in the dma-buf path
Originally, the EGL thread was extended so that it can upload images to
the GPU. This allowed CUDA and NVENC to be used for some basic hardware
accelerated encoding.
A similar path was added for dma-bufs, but since it could not be tested
at the time of implementation, it was blindly added.
In the end, it turned out, that it did not work ([0]), and thus dma-bufs
were disabled for NVIDIA setups ([1]).

Since the new way to handle hardware accelerated encoding is a zero-copy
method, simply remove the deactived handling here.

[0]: https://gitlab.gnome.org/GNOME/mutter/-/issues/3297
[1]: https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/commit/dbaa6a70dcbcd6ae457910b2483bd2f0b5d1ec76