rdp-surface-renderer: Downgrade view type of frame when necessary
When using AVC444, not every frame, that is submitted to the client,
must contain both a main view and an auxiliary view. To determine, which
views need to be submitted, it is implementation-wise easier to assume
first, that a frame needs both views. Then all cases can be checked,
whether it would make sense to downgrade the view type, i.e. not
submitting the auxiliary view.

One such check or condition is, whether the client or server side can
keep up with the frame processing, when an auxiliary frame is submitted,
as in practice, auxiliary (sub-)frames are normal frames too from the
perspective of the en- or decoder.
Whether the server side can keep up with the rendering can be easily
checked, whether there is another frame already in the render process,
when preparing the next frame.
Whether the client side can keep up is more tricky. One case, where the
client cannot keep up is when the frame controller has to throttle any
rendering operation. One other situation here is, that the frame
controller can (quickly) alternate between throttling and no-throttling.
This is not an error case. However, such situation can lead to
gnome-remote-desktop prematurely sending auxiliary frames, when the
client side did not yet fully recover yet from a situation, where
rendering in general had to be throttled. So, also account for that
situation too.

When one of those situations applies, downgrade the view type.