rdp: Also track subframe count
The previously added framerate log needs to know the current pending
stereo frame count in order to be able to make a decision, whether
another stereo frame should be avoided or not. To be able to do that,
track each frame, that contains both a main view and an auxiliary view.