rdp: Add render state helper struct
When the main view and auxiliary view are created, gnome-remote-desktop
might determine, that an auxiliary view is needed. However, there might
be a circumstance, that the auxiliary view won't be encoded, at least
for now. Such circumstance could be, that the client side might be
unable to keep up with the encoding paste of the server side.
In such situation, it is better to save the tiles, that would need an
auxiliary view, and render the auxiliary later, when the situation
improves.

To be able to communicate the affected tiles, add a render state helper
struct, which will save that information. This struct also saves the
damaged tiles. This struct will later be used by both the view creator
and the render context.