rdp-render-context: Add API to update render state
When rendering frames, there might be a situation, where a series of
frames happens, that theoretically need an auxiliary view, but
gnome-remote-desktop cannot render those, as the client would not be
able to keep up with the decoding and postprocessing. After that series,
another frame might be rendered, where that situation does not apply
anymore. This frame could pick up the missing chroma data, and
gnome-remote-desktop could encode the auxiliary view for that frame.

Use the previously added render state helper struct to update the state
of a frame. This includes updating the damage region of the frame and
updating the saved chroma data in the render context.
The saved chroma data in the render context will be used for future
frames, when the current frame does not contain an auxiliary subframe.