rdp: Update processed frames via the render context
Now, that the render state of a frame can be updated, use that API to
update processed frames. First, instead of directly creating the damage
region for a frame, create a render state, containing the damage- and
chroma-check-buffer.
The view creator will then ask the render context to update the render
state of the frame. Here, the render context might downgrade the view
type, when no auxiliary view is needed, or even extend the damage region
when picking up the saved tiles, that need extra chroma data.