rdp-gfx-frame-controller: Also keep track on current enc- and ack-rates
Each time, a frame is acked or unacked, the encoding- or
acknowledge-rate changes. Keep track of those and notify them to the
previously added framerate log.
Also, add an API to notify the frame controller, when the frame history
changed. This will be called in the next commit by the
graphics pipeline, when the frame history is rewritten. This is usually
the case, when the client suspended frame acknowledgement in the past,
but then continues with the frame acknowledgement.