rdp-renderer: Add ability to queue frame upgrades
gnome-remote-desktop can currently render frames containing a main view,
or frames containing both a main view and an auxiliary view.
However, there might be situations, where gnome-remote-desktop had to
skip the auxiliary view for a bulk of frames. Then possibly no further
frame is encoded. In that situation, there is content on the client's
screen, that needs an auxiliary view, but there is still an auxiliary
view pending.
The auxiliary view itself was already prepared, as both the main view
and the auxiliary view are created in the same shader. The auxiliary
view just needs to be encoded.

To be able to do that, extend the method to submit an RDP frame: When
the frame already has a valid view, only the encoding and submitting
operations need to be done. So, directly add the frame to the finished
view creations. This can be easily done, because in future frames, lone
auxiliary views will only be encoded, when there is no other frame left
for rendering.