rdp: Render auxiliary frames when the time is right
The renderer, as well as the RDP frame itself, are now able to handle
frame upgrades. What is missing is the preparation part in the surface
renderer.

So, add a GSource to render frame upgrades. Frame upgrades differ from
themselves from new frames, that there is no pending view creation, as
the view has already been prepared. Frame upgrades thus also don't
reference a source buffer.
Frame upgrades should not happen during a bulk of other frames for a
given surface. To ensure that, add a delay before attempting to render
an auxiliary frame. After some testing, 60ms appear to be a good value
for that.
Frame upgrades should only be rendered, when there is a pending frame
upgrade. Add an API to communicate that state to the surface renderer.
This API will be used in the next commit by the render context to update
the upgrade state.