rdp-frame: Add ability to render frame upgrades
When creating an RDP frame struct instance, the RDP frame directly knows
its source buffer. When upgrading a frame (progressively encoding a
frame), there is no new source buffer, as the current existing frame is
improved.

Use that characteristic in order to determine, whether a new frame is
about to be encoded, or whether an existing frame is improved. The
former case is already implemented. In the latter case, simply use the
previously added API to fetch the progressive render state from the
render context.
Also set the correct view type, when upgrading a frame.