rdp-render-context: Notify surface renderer of frame upgrade state
All parts needed for upgrading frames are now in place. The render
context just needs to notify the frame upgrade state to the surface
renderer.
Do this always, when the requirements are updated. That is, when either
the existence of the chroma state buffer is changed (pending extra
chroma data), or when the amount of frames, that are currently rendered
changes.