debug: Add option to debug Vulkan usage
Vulkan provides a way to debug its API usage. This allows checking
whether the API usage conforms to the documentation and gets rid of most
issues.
This check, however, is provided by an external layer, the Vulkan
validation layer, which is not installed on every system.

To only load later the validation layer and debug messenger, when
wanted, add a debug option for them.
This allows disabling the "debug mode" even when the validation layer is
available.