rdp: Add way to try acquiring render-context without recreating it
The render context saves the progressive render state. When it is
deleted, some substantial change is happening, and the encode session
needs to be recreated. So, upgrading a frame is only possible, when the
render context is not preserved. The surface renderer, however, has no
influence on, whether the render context is preserved or not. This is
also not possible, because the RDP client can force a protocol reset of
the graphics pipeline at any time, or submit a new monitor config at any
time, leading to a change of the monitor layout.

In order to be able to acquire the render context safely for frame
upgrade operations, add a flag, that asks the renderer to preserve the
render context or return NULL.
When this flag is used, the acquired render context will always contain
the progressive render state (also because new render contexts don't
have a progressive render state yet). If the render context could not be
acquired, rendering is then either suspended, inhibited, or the client
invoked a change (e.g. new monitor config), that led to a reset of the
render context.