rdp-render-context: Add API to fetch progressive render state
In future commits, render handling of auxiliary-view-only frames is
added. When such rendering occurs, there won't be any other frame for
the given surface, that is rendered in parallel.
When the RDP frame acquires the image views needed for a render
operation, it will always acquire the image view for the auxiliary view
as last from the render context.
The render context also knows the region, that is in need of the missing
extra chroma data.

So, add an API to the render context to fetch both items. These will be
used in future commits to progressively improve the last frame.