rdp: Add class for tracking framerates
Currently, gnome-remote-desktop only encodes an auxiliary view, when
there is a need of extra chroma data for a frame, when the server side
can keep up, and when the client side can keep up.
However, the mechanism to detect, whether the client can keep up depends
on the past history of frames.
gnome-remote-desktop cannot fully predict the situation in a given
moment. This is usually the case, when gnome-remote-desktop prepares a
bulk of stereo frames, while the frame controller only knows the result
of the past frames.

As a result, the client gets bombarded with multiple stereo frames,
which require more time in decoding and postprocessing, than usual
frames. This is visible, when playing a video, where the client side can
keep up with frames, containing only the main view, but cannot keep up
with a bulk of frames, containing stereo views.
In that case, gnome-remote-desktop's encoding-rate is stable, but the
clients frame-acknowledge-rate behaves like a sinus curve, while the
watched video visibly stutters.
To take care of this situation, gnome-remote-desktop needs to be able to
detect this situation and degrade the frame view type in those
situations.
In order to do that, keep a log of the past encoding rates. If they are
above a certain value, or are stable with respect to their value
differences, then evaluate the situation as "watching a video". If the
frame acknowledge rate is now lower than the encoding rate, degrade the
frame view type. Usually, when stereo frames are rendered, they come in
a bulk, so also account for that too, when determining whether to
degrade the frame view type.
This handling still allows stereo frames during videos, for example,
when a video does not fill the whole screen, but only a portion of it.
Surrounding (static) content, that requires extra chroma data, can then
still be improved without impacting the framerate of the played video.