rdp: Add API to check whether stereo frames should be avoided
This allows the surface renderer in the next commit to determine,
whether the auxiliary view of the next frame should not be encoded.
Also, use the very same API to degrade the view type of an already
preprocessed frame. This allows gnome-remote-desktop to adapt itself
more quickly to a situation, where the framerate of the client side
drops, while the encoding rate does not.