Add class for conveniently handling a Vulkan memory object
The creation of buffer objects and image objects in Vulkan is done in an
explicit way.
This means, that first the buffer or image is created, then the memory
requirements for that buffer or image are acquired, then a memory object
is created, and then the buffer or image is bound to that memory object.

For convenice, add a GObject class, which will maintain a VkMemory
object.
This class is also capable of importing memory from dma-buf- or
opaque-fds in addition to map the memory object, so it can be accessed
in host memory.