Add class for conveniently handling a Vulkan buffer object
Now, with the convenience class for VkMemory objects being added, add a
convenience class for Vulkan buffer objects.
This class will later be used in compute shaders to save for example the
damage detection state (whether a tile is damaged).
Upon initialization of this class, create and bind a respective VkMemory
object.