Add class for conveniently handling a Vulkan image object
Similarly to the previous class for Vulkan buffer objects, add a class
for Vulkan image objects.
The primary objective for this class is to maintain imported dma-buf
images for later usage.
Creating an image object in Vulkan is more complex and requires that the
API user checks whether the device supports the image with its wanted
attributes. Intels Vulkan driver is very strict here: If the driver does
not support the wanted image, it will crash the whole application.

In order to be able to access Vulkan image objects later, a VkImageView
is needed. So, automatically create one too, when creating a VkImage.