Add helper function to easily extend Vulkan function calls
Most Vulkan structs used in Vulkan function calls are extensible. What
this means is that a function does a certain thing, e.g. retrieving
information about a device and saving the information in some struct.
And to this struct another struct can be appended, where additional
information is stored.
Usually, these additional structs are provided by extensions.

For this, these extensible structs first contain an enum value for their
type and a pointer to the next struct.
Extending a struct and with which structs depends on the use case and
device. And so, it makes it easier to not know, what is already in a
chain of structs (otherwise there would be a ton additional
if-statements later).

To make the Vulkan API usage easier, add a method, which adds a struct
to the chain of another struct. This API will be used in lots of
different places later.
Also, since this is the first Vulkan usage here, add a dependency to
Vulkan itself.