rdp-frame: Acquire required image views on initialization
When trying to create a frame (struct), also acquire the image views for
view creator.
The view creator will later use those to create (fill with content)
those views, and the encode session will alter use those to encode the
actual subframes of the frame.
For now, use just one image view for AVC444. This makes the AVC
implementation easier. In the future, the image view count for AVC444
will be set to two views, when actual AVC444 handling is implemented.