rdp-renderer: Add render handling needed for zero-copy rendering
With all pieces for the renderer now in place, add the render handling,
which will allow zero-copy rendering. The idea of the renderer is the
following:
When using hardware acceleration, some actions can be parallelized. For
example, when having an AVC encode session, then first the image views
need to be prepared (colour conversion, creation of the auxiliary view),
then the actual encode operation(s) happen(s), and then the frame is
submitted via the graphics pipeline.
The first action will be done by the shader units on the GPU, the second
action by the media chip on the GPU, the third one is done by the CPU.

All those three actions use different resources, and this allows
gnome-remote-desktop to run them in parallel. The advantage of this is,
that the actual framerate is improved. GPUs have one downside: When
submitting a task to them, they won't immediately start them. There are
multiple reasons for why this can happen. However, regardless of the
reason, when all parallelizable subtasks for a frame render operation
are parallelized, then this issue is marginalized, as then the framerate
is not dependent on the length of all subtasks, but instead on the
largest subtask of all subtasks.
The frame latency remains the same, however, sudden spikes in 'pending'
state of an operation will less likely become visible to the remote
desktop user.

A frame in the renderer has a maximum of four stages:
1. The first stage is the 'queued' stage. The surface renderer submitted
   a frame, but the renderer hasn't picked it up yet. The renderer will
   try to pick this frame up, when no pending frame is in stage two.
   Any frame, that is in stage one, can be replaced by the surface
   renderer, if there is a newer PipeWire buffer available.
2. In stage two, the view is created. For AVC, this means, that the
   colour conversion happens, the damage detection is done and later
   also a check happens, whether an auxiliary view is actually needed.
   If the frame has no damage, it will be dropped after this stage.
3. In stage three, all views associated with a frame are encoded.
4. Stage four is the last stage. In this stage, the frame is submitted
   to the client. When the RDP8 bulk compression is implemented in
   FreeRDP, then gnome-remote-desktop will be able to further save some
   bits here.

Additionally to the render handling itself, also add a function to be
able to submit frames to the renderer. This API will be used by the
surface renderer in future commits.