rdp-surface-renderer: Add dummy methods to seemingly invalidate surface
When adding the new render path, which will allow rendering of surfaces
in a parallel matter, a new problem will rise: When invalidating all
surfaces, a deadlock can happen. Invalidating all surfaces happens in
two cases:

1. The graphics pipeline is being reset
2. A new monitor config sent by the client was applied

The first case will work just fine. However, in the second case, the
surface renderer will likely still hold the PW buffer of the last frame
(the handling for that is added in future commits), that was rendered in
order to be able to do damage comparison.
When the monitor handling changed, however, the last frame needs to be
dropped. To do that without race conditions, a mutex is needed, which
will be locked in case 1.
In case 2, the surface invalidation will happen in a function call,
where the mutex is already locked.

To be able to handle both cases, add methods to invalidate the surface,
when the respective render mutex is already locked, and when the
respective render mutex is not yet locked.

The invalidation handling itself will be added later. It cannot be added
right now due to a circular dependency in the code.
So for now, just add dummy implementations for the surface invalidation
handling.