Add VAAPI AVC encode session
With the base class for encode sessions and the new render handling
being implemented, add the AVC encode session using VAAPI.
For the source surfaces, that will contain the prepared subsampled image
views, use a hash table.
These surfaces are here called NV12VkVASurface. These NV12VkVASurface
contain the VA surface, used by libva, and the corresponding image view,
used the view creator later.
The VA surface is created first, then exported, so that it can be
imported as dma-buf into Vulkan.

When creating the VAAPI AVC session, first check whether the stream can
actually be encoded with H.264, as for an H.264 stream a so-called
level-idc value needs to be specified. This value depends on both the
surface size, as well as on the framerate. This value is also used by
the client side to check, whether a hardware decoder can be used.
Also check the surface constraints. The GPU might be able to encode
frames into a H.264 stream, but might not support every surface size.
After both checks complete successfully, the encode session can be
created.

For each frame, introduce a new struct called H264Frame. An instance of
this struct is created for every frame, that is about to be encoded.
This frame contains relevant information for that frame, for example the
source surface, the reference picture, the NAL bitstream headers, and
the pointer to the bitstream buffer for the AVC frame.

When the frame is submitted for encoding, add it to a hash table. When
adding support for AVC444 later, multiple frames will be queued for
encoding. The encode session (then) needs to track each frame.

When an encoding operation is done, do not duplicate the bitstream. When
hardware acceleration is used for encoding, gnome-remote-desktop spends
most of its time in copying memory. While the bitstream is smaller than
the source frame, it still usually takes more than 100µs to copy that
bitstream. Save that time by just using a reference.
The bitstream is then locked (the term comes from NVENC), and then
unlocked, when the bitstream becomes unused.
Unfortunately, libva does not seem to support to reuse a bitstream
buffer, so recreate one for every frame, that is encoded.