Add image-view for RGB images
This image-view will be used for RFX encode-sessions. Like all
encode-sessions, the RFX encode-session will allocate the image-views.
The difference to NV12 image-views is here, that the image-view for RGB
images allows a local buffer to be attached to the image view.
As mentioned in previous commits, this is needed, as the colour
conversion is done by the RFX encode-session itself and thus the
underlying RDP buffer is still needed after view creation.