Add encode-context helper struct
Currently, encode-sessions encode frames all in the same way (same
quantizer parameters, same damage region). For VA-VAPI, this is not
really an issue, as the driver itself determines the damage region.
However, other encoders, like the FreeRDP's RFX software encoder need
the damage region to be passed to them. Additionally, in the future,
gnome-remote-desktop might need to render a frame at lower quality when
the available bandwidth is not enough.

To take care of these issues, introduce an encode-context, which
contains additional frame specific encode parameters. For now, this is
just the damage region, as that is what the RFX encode-session will
need.