rdp: Add class for RFX software encoding
Encoding a frame with the software-based implementation of the RFX codec
in FreeRDP makes use of all CPU threads. Thus, it does not make sense to
run multiple parallel RFX encodings at the same time.
Each RDP render context has with the new render path though its own
encode session. In case of RFX software encoding needs to be used, some
mechanism needs to be introduced to ensure, that only one software
encoding operation happens at the same time.

To do this, add a class that encodes a frame with the RFX codec. The
instance of this session is then shared among all RFX software-based
encode sessions. Then add a mutex inside this class to ensure, that only
one encode operation happens at the same time.