Pass encode-context to encode-sessions
The encode-session is remote-protocol unspecific, so the RDP frame is
not directly passed to the encode-session. So, retrieve the
encode-context from the RDP frame and pass it to the encode-session,
when doing the encode_frame() call.
This allows the encode-session to use frame specific parameters, such as
the damage region, when encoding this specific frame.