image-view: Add API to notify an image-view release
When preparing the view for a frame, the current and last RDP buffers
are passed to the view creator. For AVC, the view creator uses these
buffers to perform the colour conversion and damage detection. After
this, both buffers won't be needed anymore, as the view now carries the
data needed for the encode session.
With RFX (Calista) Progressive, this is different. FreeRDPs RFX software
encoder performs the colour conversion itself, thus the current RDP
buffer will still be needed after view creation. This is at least the
case for MemFd-based encode-sessions.
Additionally, there is a difference between MemFd-based encode-sessions
and dma-buf-based encode-sessions:
For dma-buf-based encode-sessions, the image content needs to be
downloaded to a local buffer, so that the RFX encode-session can encode
the frame. In MemFd-based encode-sessions, the buffer contents are
already locally available and thus don't need to be copied.
Both situations need to be somehow handled.

To take care of this issue, an abstract local buffer class is needed for
the RFX encode-session. For dma-buf-based encode-sessions, this class
will be extended to carry a copy of the dma-buf image. For MemFd-based
encode-sessions, this class will be extended to hold the actual RDP
buffer underneath it.
These buffers are needed for the encode-sessions, but the image-view is
the one actually passed to the encode-session. So, for BGRX images these
local buffers need to be attached to the image view. When the image-view
is not needed anymore, the local buffer won't be needed anymore too.
This needs to be communicated to the origin, where the local buffer is
claimed, so that it can be released again. The encode-session won't be
that origin, as it can and should be independent from the actual
view-creators.
So, add an API to notify this image-view release. The other parts are
added in the next commits. Also add a dummy implementation to the
image-view for NV12 images, as the intention is to notify the image-view
release unconditionally.