rdp: Add generic view-creator using GL
When software encoding is used, every source buffer, that is used for
encoding needs to lay in host-memory. So, this view-creator basically
downloads any dma-buf image and performs the damage detection.
Since the RFX encode-session performs the colour conversion itself,
there is nothing else that needs to be done.
Each local buffer, in which the dma-buf image is downloaded, needs to be
created by the view-creator, as the encode-session is intended to be
used by both the dma-buf and MemFd path and thus the encode-session does
not get to know the local-buffer implementation, but the local-buffer
abstraction instead.

When downloading the dma-buf image, a local-buffer is acquired first.
When the image is downloaded, the damage detection is done using the now
downloaded dma-buf image, and - if available - the downloaded dma-buf
image, that was downloaded before the current image.
After the damage detection handling, the current buffer is set as the
old buffer. So, for this handling here, two local buffers are needed.
However, a third one is needed:
An acquired local-buffer is released, when the associated RDP frame is
freed. That is usually the case after it was sent to the client.
When the surface-render-source in the RDP renderer runs, there can the
following situation in one run:
- The last encoding is done and the frame is ready to be submitted to
  the client. Here, the source buffer (local-buffer) is still claimed.
- The last view creation is also done, and the frame is submitted to the
  encode-session. The local source buffer here is also claimed.
- The next PW buffer is ready, the renderer submits the associated RDP
  buffer to the view-creator, a local-buffer for the dma-buf image
  download is acquired (claimed).

So, in total, allocate three local-buffers to account for such
situations.