rdp-renderer: Create fallback software encoder when starting session
This encoder will be used by every software-based RFX (Calista)
encode-session to do the actual encoding. This ensures, that only one
encoding, that uses all CPU threads, happens at the same time.
The reference of this encoder will in the next commit be passed to the
render context, where the encode-session is created.