egl-thread: Add API to get the DRM render node of the device
Vulkan does not know which GPU is used by the wayland compositor, nor
does it provide a way to get that information.
EGL, however, does know that, since EGL can check whether the used
platform is the wayland platform.

For interoperability, both Vulkan and GL have a way to ensure, that the
same GPU is used.
This method can be used as a workaround for the aforementioned problem:
First, initialize the EGL thread, and then initialize Vulkan with the
very same GPU.
Checking, that the same GPU is used is done by comparing the DRM render
nodes.

Add a method to the EGL thread retrieve the DRM render node allowing
Vulkan to match the same GPU in later commits.
For simplicity, directly fetch the DRM render node from EGL, when the
EGL thread is initialized.