debug: Add option for the new render handling
The new render handling is not complete yet, as the fallback handling,
when no AVC or dma-bufs are available, as well as the explicit sync
support is not there yet.
This debug option allows adding support for the zero-copy render
handling via Vulkan and VAAPI without breaking setups.
When the fallback renderer and support for explicit sync is added, this
option won't be needed anymore.