session-rdp: Use submitted keyboard type, instead of hardcoding it
The RDP scancodes depend on the keyboard type, that the client side
submitted. Currently, gnome-remote-desktop hardcodes the keyboard type
to IBM enhanced, but this is not always true, like in the case of the
Japanese keyboard.

Fix this issue by actually using the keyboard type, that the client side
submitted, instead of hardcoding the IBM enhanced keyboard type.