rdp: Rename GrdRdpDvc to GrdRdpDvcHandler
In order to be able to abstract all dynamic virtual channels, a proper
name for the base class is needed. GrdRdpDvc fits perfectly for this.
However, that name is already used for the DVC handler in order to
notify DVCs, when they have been created protocol-wise.
So rename GrdRdpDvc to GrdRdpDvcHandler in order to free up that name.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/313>