rdp: Add base class for DVCs
This base class is used to abstract all classes for DVCs from it. Add
methods to subscribe and unsubscribe the respective DVC status, as this
operation is done by all DVC implementations inside
gnome-remote-desktop.
Also add a vfunc to trigger the initialization of the DVC. This makes
abstracting the initialization possible in order to reduce duplicated
code in the future.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/313>