session-rdp: Add graphics pipeline to RDP channel enum
The graphics pipeline is the only DVC in gnome-remote-desktop, that is
not part of the GrdRdpChannel enum. In order to be able to move further
commonly used DVC functionality to the DVC base class, add the graphics
pipeline to the GrdRdpChannel enum.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/313>