rdp-dvc: Add method to tear down channel
All DVC implementations in gnome-remote-desktop, except for the graphics
pipeline, are optional. Thus, when such DVC cannot be opened, it will be
torn down, as it is unneeded. The process for this is always the same.
So add a method to the base DVC class to tear down the respective DVC
implementation. This method will replace all other existing
implementations of the DVC teardown in the next commits, effectively
removing duplicated code.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/313>