rdp/rdpgfx: Don't try to open DVC again when opening it already failed
When opening the graphics pipeline DVC fails, the whole session is torn
down, as the graphics pipeline is essential for gnome-remote-desktop.
However, there is currently the possibility, that gnome-remote-desktop
tries to open that DVC again, despite the session already being in the
process of being stopped.
So add a helper variable to prevent that.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/313>